first push
ability to compile using visual studio 2022
disabled deprecation warnings
solved a few compilation errors by doing code modifications
still requires the original old dependencies
still 32bits

+++ sources/impl/MeshLoaderCollada.cpp
@@ -2130,7 +2130,7 @@ namespace hpl {
 			{
 				//Get number of chars in prefix.
 				int lPrefixChars =1;
-				while(	lPrefixChars < apNode->msName.size() &&
+				while(	lPrefixChars < (int)apNode->msName.size() &&
 						apNode->msName[lPrefixChars]!= '_' &&
 						apNode->msName[lPrefixChars]!='\0') {
 						lPrefixChars++;
@@ -2360,9 +2360,9 @@ namespace hpl {
 
 				//Get number of digits
 				lDigits =1;
-				while(	apNode->msName.length() > lStartChar+4+lDigits &&
+				while(	(int)apNode->msName.length() > lStartChar+4+lDigits &&
 						apNode->msName[lStartChar+4+lDigits] != '_' &&
-						apNode->msName.length() >= 7+lDigits &&
+						(int)apNode->msName.length() >= 7+lDigits &&
 						apNode->msName[7+lDigits] != 0)
 				{
 					lDigits++;
